\name{DectechR-package}
\alias{DectechR-package}
\alias{DectechR}
\docType{package}
\title{
Useful functions for Dectech analysts
}
\description{
Collection of functions useful for Dectech analysts, such as getting extracting survey data from Questback output, and getting consistent coefficient tables from regressions.
}
\details{
\tabular{ll}{
Package: \tab DectechR\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2016-02-09\cr
License: \tab GPL-2\cr
}
Useful functions to have loaded during general data analysis at Dectech.
}
\author{
Keith Simpson

Maintainer: Keith Simpson <k.simpson@dectech.co.uk>
}


\keyword{ package }

\examples{
  x = c(1,2,3)
  cc(x)
}
